+++ vega_datasets/__init__.py
@@ -3,7 +3,3 @@ from vega_datasets.core import DataLoader, LocalDataLoader
 data = DataLoader()
 local_data = LocalDataLoader()
 __version__ = "0.9.0dev0"
-
-# This is the tag in http://github.com/vega/vega-datasets from
-# which the datasets in this repository are sourced.
-SOURCE_TAG = "v1.29.0"