+++ vega_datasets/core.py
@@ -5,8 +5,11 @@ import pkgutil
 import textwrap
 from typing import Any, Dict, Iterable, List
 from urllib.request import urlopen
-
 import pandas as pd
+
+# This is the tag in http://github.com/vega/vega-datasets from
+# which the datasets in this repository are sourced.
+SOURCE_TAG = "v1.29.0"
 
 
 def _load_dataset_info() -> Dict[str, Dict[str, Any]]:
@@ -91,7 +94,7 @@ class Dataset(object):
     _reference_info = """
     For information on this dataset, see https://github.com/vega/vega-datasets/
     """
-    base_url = "https://vega.github.io/vega-datasets/data/"
+    base_url = "https://cdn.jsdelivr.net/npm/vega-datasets@" + SOURCE_TAG + "/data/"
     _dataset_info = _load_dataset_info()
     _pd_read_kwds = {}  # type: Dict[str, Any]
     _return_type = pd.DataFrame